+++ Source/Scene/SceneTransitioner.js
@@ -90,10 +90,10 @@ define([
 
         this._scene = scene;
         this._ellipsoid = defaultValue(ellipsoid, Ellipsoid.WGS84);
-        var context = scene.getContext();
+        var context = scene.context;
 
         // Position camera and size frustum so the entire 2D map is visible
-        var maxRadii = this._ellipsoid.getMaximumRadius();
+        var maxRadii = this._ellipsoid.maximumRadius;
         var position = new Cartesian3(0.0, 0.0, 2.0 * maxRadii);
         var direction = Cartesian3.normalize(Cartesian3.negate(position));
         var up = Cartesian3.UNIT_Y;
@@ -368,7 +368,7 @@ define([
 
     function createMorphHandler(transitioner, completeMorphFunction) {
         if (transitioner.completeMorphOnUserInput) {
-            transitioner._morphHandler = new ScreenSpaceEventHandler(transitioner._scene.getCanvas());
+            transitioner._morphHandler = new ScreenSpaceEventHandler(transitioner._scene.canvas);
 
             var completeMorph = function() {
                 transitioner._morphCancelled = true;
@@ -384,7 +384,7 @@ define([
     }
 
     function destroyMorphHandler(transitioner) {
-        var animations = transitioner._scene.getAnimations();
+        var animations = transitioner._scene.animations;
         for ( var i = 0; i < transitioner._currentAnimations.length; ++i) {
             animations.remove(transitioner._currentAnimations[i]);
         }
@@ -395,14 +395,14 @@ define([
     function morphFromColumbusViewTo3D(transitioner, duration, onComplete) {
         var scene = transitioner._scene;
 
-        var camera = scene.getCamera();
+        var camera = scene.camera;
         setCameraTransform(camera, Matrix4.IDENTITY);
 
         var startPos = camera.position;
         var startDir = camera.direction;
         var startUp = camera.up;
 
-        var maxRadii = transitioner._ellipsoid.getMaximumRadius();
+        var maxRadii = transitioner._ellipsoid.maximumRadius;
         var endPos = transitioner._ellipsoid.cartographicToCartesian(new Cartographic(0.0, 0.0, 10.0));
         endPos = Cartesian3.multiplyByScalar(Cartesian3.normalize(endPos), 2.0 * maxRadii);
         var endDir = Cartesian3.normalize(Cartesian3.subtract(Cartesian3.ZERO, endPos));
@@ -416,7 +416,7 @@ define([
             camera.right = Cartesian3.cross(camera.direction, camera.up);
         };
 
-        var animation = scene.getAnimations().add({
+        var animation = scene.animations.add({
             duration : duration,
             easingFunction : Tween.Easing.Quartic.Out,
             startValue : {
@@ -435,7 +435,7 @@ define([
     function morphFrom2DTo3D(transitioner, duration, onComplete) {
         duration = duration * 0.5;
 
-        var camera = transitioner._scene.getCamera();
+        var camera = transitioner._scene.camera;
 
         morphOrthographicToPerspective(transitioner, duration, function() {
             camera.frustum = transitioner._cameraCV.frustum.clone();
@@ -451,7 +451,7 @@ define([
 
     function morphPerspectiveToOrthographic(transitioner, duration, onComplete) {
         var scene = transitioner._scene;
-        var camera = scene.getCamera();
+        var camera = scene.camera;
 
         var startPos = camera.position;
         var startFOVy = camera.frustum.fovy;
@@ -466,7 +466,7 @@ define([
             camera.position = Cartesian3.multiplyByScalar(Cartesian3.normalize(camera.position), distance);
         };
 
-        var animation = scene.getAnimations().add({
+        var animation = scene.animations.add({
             duration : duration,
             easingFunction : Tween.Easing.Quartic.Out,
             startValue : {
@@ -486,8 +486,8 @@ define([
 
     function morphFromColumbusViewTo2D(transitioner, duration, onComplete) {
         var scene = transitioner._scene;
-        var camera = scene.getCamera();
-        var maxRadii = transitioner._ellipsoid.getMaximumRadius();
+        var camera = scene.camera;
+        var maxRadii = transitioner._ellipsoid.maximumRadius;
 
         setCameraTransform(camera, transitioner._cameraCV.transform);
 
@@ -511,7 +511,7 @@ define([
         };
 
         duration = duration * 0.5;
-        var animation = scene.getAnimations().add({
+        var animation = scene.animations.add({
             duration : duration,
             easingFunction : Tween.Easing.Quartic.Out,
             startValue : {
@@ -531,7 +531,7 @@ define([
     function morphFrom3DTo2D(transitioner, duration, onComplete) {
         duration = duration * 0.5;
 
-        var maxRadii = transitioner._ellipsoid.getMaximumRadius();
+        var maxRadii = transitioner._ellipsoid.maximumRadius;
 
         var tanPhi = Math.tan(transitioner._camera3D.frustum.fovy * 0.5);
         var tanTheta = transitioner._camera3D.frustum.aspectRatio * tanPhi;
@@ -550,8 +550,8 @@ define([
 
     function morphOrthographicToPerspective(transitioner, duration, onComplete) {
         var scene = transitioner._scene;
-        var camera = scene.getCamera();
-        var maxRadii = transitioner._ellipsoid.getMaximumRadius();
+        var camera = scene.camera;
+        var maxRadii = transitioner._ellipsoid.maximumRadius;
 
         var tanPhi = Math.tan(transitioner._cameraCV.frustum.fovy * 0.5);
         var tanTheta = transitioner._cameraCV.frustum.aspectRatio * tanPhi;
@@ -598,7 +598,7 @@ define([
             }
         }
 
-        var animation = scene.getAnimations().add({
+        var animation = scene.animations.add({
             easingFunction : Tween.Easing.Quartic.Out,
             duration : partialDuration,
             startValue : {
@@ -618,7 +618,7 @@ define([
 
     function morphFrom2DToColumbusView(transitioner, duration, onComplete) {
         var scene = transitioner._scene;
-        var camera = scene.getCamera();
+        var camera = scene.camera;
 
         duration = duration * 0.5;
 
@@ -638,7 +638,7 @@ define([
                 camera.right = Cartesian3.cross(camera.direction, camera.up);
             };
 
-            var animation = scene.getAnimations().add({
+            var animation = scene.animations.add({
                 duration : duration,
                 easingFunction : Tween.Easing.Quartic.Out,
                 startValue : {
@@ -662,7 +662,7 @@ define([
     function morphFrom3DToColumbusView(transitioner, duration, endCamera, onComplete) {
         var scene = transitioner._scene;
 
-        var camera = scene.getCamera();
+        var camera = scene.camera;
         setCameraTransform(camera, transitioner._cameraCV.transform);
 
         var startPos = Cartesian3.clone(camera.position);
@@ -680,7 +680,7 @@ define([
             camera.right = Cartesian3.cross(camera.direction, camera.up);
         };
 
-        var animation = scene.getAnimations().add({
+        var animation = scene.animations.add({
             duration : duration,
             easingFunction : Tween.Easing.Quartic.Out,
             startValue : {
@@ -715,14 +715,14 @@ define([
             };
         }
 
-        var animation = scene.getAnimations().addProperty(scene, 'morphTime', start, stop, template);
+        var animation = scene.animations.addProperty(scene, 'morphTime', start, stop, template);
         transitioner._currentAnimations.push(animation);
     }
 
     function updateFrustums(transitioner) {
         var scene = transitioner._scene;
 
-        var context = scene.getContext();
+        var context = scene.context;
         var ratio = context.getDrawingBufferHeight() / context.getDrawingBufferWidth();
 
         var frustum = transitioner._camera2D.frustum;
@@ -737,7 +737,7 @@ define([
         frustum = transitioner._camera3D.frustum;
         frustum.aspectRatio = ratio;
 
-        var camera = scene.getCamera();
+        var camera = scene.camera;
         switch (scene.mode) {
         case SceneMode.SCENE3D:
             camera.frustum = transitioner._camera3D.frustum.clone();
@@ -759,7 +759,7 @@ define([
         destroyMorphHandler(transitioner);
 
         updateFrustums(transitioner);
-        var camera = scene.getCamera();
+        var camera = scene.camera;
         camera.transform = Matrix4.clone(Matrix4.IDENTITY);
 
         if (transitioner._previousMode !== SceneMode.MORPHING || transitioner._morphCancelled) {
@@ -786,7 +786,7 @@ define([
         destroyMorphHandler(transitioner);
 
         updateFrustums(transitioner);
-        var camera = scene.getCamera();
+        var camera = scene.camera;
         camera.transform = Matrix4.clone(transitioner._camera2D.transform);
 
         // TODO: Match incoming columbus-view or 3D position
@@ -808,7 +808,7 @@ define([
         destroyMorphHandler(transitioner);
 
         updateFrustums(transitioner);
-        var camera = scene.getCamera();
+        var camera = scene.camera;
         camera.transform = Matrix4.clone(transitioner._cameraCV.transform);
 
         if (transitioner._previousModeMode !== SceneMode.MORPHING || transitioner._morphCancelled) {