+++ Source/Scene/ScreenSpaceCameraController.js
@@ -247,7 +247,7 @@ define([
         this._horizontalRotationAxis = undefined;
 
         // Constants, Make any of these public?
-        var radius = this._ellipsoid.getMaximumRadius();
+        var radius = this._ellipsoid.maximumRadius;
         this._zoomFactor = 5.0;
         this._rotateFactor = 1.0 / radius;
         this._rotateRateRangeAdjustment = radius;
@@ -274,7 +274,7 @@ define([
      */
     ScreenSpaceCameraController.prototype.setEllipsoid = function(ellipsoid) {
         ellipsoid = ellipsoid || Ellipsoid.WGS84;
-        var radius = ellipsoid.getMaximumRadius();
+        var radius = ellipsoid.maximumRadius;
         this._ellipsoid = ellipsoid;
         this._rotateFactor = 1.0 / radius;
         this._rotateRateRangeAdjustment = radius;